requirements.txt
telethon==1.42.0
python-dotenv==1.2.1
qrcode==8.2.0